agentic-ai/backend/app/rag/requirements.txt
openai==2.3.0
python-dotenv>=1.0.0

# Retrieval stack
sentence-transformers>=2.5.1,<3.0
faiss-cpu>=1.7.4,<2.0
rank-bm25==0.2.2
tiktoken>=0.5.2

# Utils
numpy>=1.24,<2.0
rich>=13.7